+++ lib/src/extensions/duration.dart
@@ -1,6 +1,6 @@
 part of './../extensions.dart';
 
-extension DurationExtensions on Duration {
+extension MayrDurationExtensions on Duration {
   Future delay([FutureOr Function()? callback]) async =>
       Future.delayed(this, callback);
 }