+++ lib/src/extensions/stateless_widget.dart
@@ -1,3 +1,3 @@
 part of './../extensions.dart';
 
-extension CustomStatelessWidget on StatelessWidget {}
+extension MayrStatelessWidgetExtensions on StatelessWidget {}