+++ lib/src/extensions/text_style.dart
@@ -1,3 +1,3 @@
 part of './../extensions.dart';
 
-extension CustomTextStyle on TextStyle {}
+extension MayrTextStyleExtensions on TextStyle {}