+++ lib/src/extensions/text_widget.dart
@@ -1,3 +1,3 @@
 part of './../extensions.dart';
 
-extension CustomTextWidget on Text {}
+extension MayrTextWidgetExtensions on Text {}