+++ test/extensions/dynamic_test.dart
@@ -1,0 +1,9 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:mayr_extensions/mayr_extensions.dart';
+
+void main() {
+  test("orDefault", () {
+    int? number;
+    expect(number.orDefault(1), 1);
+  });
+}